zalo_chatlog_oos_negative_pipeline/src/requirements.txt
black~=22.0
fastparquet==2023.4.0
flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0
isort~=5.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0
jupyter~=1.0
kedro-datasets[pandas.ParquetDataset]
kedro-telemetry~=0.2.0
kedro-viz==7.1.0
kedro~=0.18.4
matplotlib==3.7.3
nbstripout~=0.4
numpy==1.22.4
openpyxl==3.1.2
pandas==1.5.2
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=6.2
seaborn==0.12.2
tqdm==4.66.1
omegaconf==2.3.0
